Add support for the publication of the list of all published objects.
This is needed by the CCDB backend as we have no ways of asking for the
list of objects. It introduces the (first) so-called system object, i.e.
an object that is transmitted by tasks behind the scene the same way
the explicit MonitorObjects are.
At the moment, the list of objects is a comma separated list of names.
It is interpreted by the CCDB backend to provide the list of published
objects.
(QC-53)