Information Service  (#6)
InformationService to broadcast list of tasks and their objects

Add an information service that collects the list of objects published by all the tasks and make it available to clients. There are 2 binaries : qcInfoService and qcInfoServiceDump. The latter is to check what is being produced by the Information Service.

The information service can use a file as input if there are no tasks running but one wants to simulate a normal use.
QC-57