Reworked QC infrastructure configuration and generation
This commit introduces a main qc factory which can be used to generate
a full topology with just one function call. As the TaskRunner can be
placed either locally on FLP or EPN machine or remotely on dedicated
QC servers, a distinction between this two has been introduced. When a task
is 'local', the taskRunners will be generated by generateLocalInfrastructure(),
but mergers and checkers by generateRemoteInfrastructe(). If a task is
'remote', both taskRunner and checker will be generated as remote
infrastructure.

The commit removes also some deprecated config files.