Pass --run to testWorkflow
The DPL driver, when followed by |, think it should dump the workflow
instead of running it. We need to pass --run option to enforce it.
add_test takes a list of arguments and puts each of them inside quotation
marks, so i had explicilty to split "-b --run" into "-b" and "--run".