[EMCAL-524] Prototype QC component for EMCAL digits (#198)
Digits component monitoring digit amplitude and time
for each tower separately. Creating also the EMCAL
module.

In order to run the Digits QC task a separate workflow
(qcRunEmcalDigit) is used. The data source is an
EMCAL publisher implemented in the O2 EMCAL reco
workflow supporting only digits for the moment. The
QC task are initialized using the standard method
via a configuration file.