Syncing to new DPL CompletionPolicy API after O2 PR 2886
Pull request AliceO2Group/AliceO2#2886 changes the argument of the policy
callback from span<PartRef> to span<DataRef> which is a span of non-message-owning
reference objects.

In order to make depending packages easier to implement, the callback signature is
now defined in CompletionPolicy.h header in AliceO2Group/AliceO2#2918, thee
definitions are now used in the TaskRunner.