[EMCAL-526] Convert online to offline mapping (#756)
* [EMCAL-526] Convert online to offline mapping

Raw row/col IDs in the DCAL are different in the
online and offline scheme. As the cell index mapping
is using the offline scheme, using the online scheme
will lead to a crash when receiving data from SM 19
(DCAL 1/3) due to invalid cell ID. Local co and row
need to be shifted back to offline index scheme using
the corresponding wrapper function in the geometry.

In addition refactor jfecID -> supermoduleID for
readability.

* [EMCAL-526] fix reading of FECid

* [EMCAL-526] Use input record walker in order to scan all links

Old style iteration (copied from DAQ task) was only
processing the first link. Code is taked from the
Raw-to-cell converter.

* [EMCAL-526] Fix labels for calib and physics trigger

Labels were swapped with respect to the event type
used in the histograms.

Co-authored-by: Cristina Terrevoli <cristina.terrevoli@cern.ch>