[EMCAL-704] Add suport of cell container in multiple subtimeframes
In case the reconstruction runs on the FLP the cells
might come from two different cell containers from
the two subtimeframes. In this case the cell containers
might not be merged, but instead the TriggerRecords
are merged to have two different RangeReferences
with a subspecification, called Subevent. The event
loop iterates over the CombinedEvent containing
the interaction record, the trigger type and the list of
subevents. In order to identify the cell container based
on the subspecification a map is built at the beginning
of the monitorData function.