[EMCAL-524] Fix handling of event building in case of downsampling
Cell and trigger record container cannot be filtered
according to the channel type as it changes in case
the a data sampling policy is applied. Also for
wildcarded inputs the iteration must be done but
based on the binding. As no matcher exist for the
binding the iteration is done by hand.