[MCH] fixed memory leak in custom mergeable histograms (#1022)
The memory leak was due to the fact that internal histograms of the
mergeable objects were not deleted in the object's destructor.

The code also reduces the number of objects published by default by the
digits task. A "Diagnostic" custom parameter re-enables the publishing of
the additional histograms (for experts only).