[MCH] added processing of DPL message with errors vector
In addition to the direct processing of the raw data, this commit
adds the possibility to process and display a vector of errors received
in a dedicated DPL message.

This functionality is intended to provide a centralized task that handles
the errors coming from different stages of the O2 processing.