[QC-938] PP tasks may have the same names among different detectors
Also, it is ensured that a QC task and PP task may have the same name for one detector.

This decouples ID and name similarly as for QC tasks, checks and aggregators.
Thus, ID has to be unique and is the key of the pp task config structure.
Then, "taskName" can be specified inside the structure.
If it is empty, taskID is used as taskName.