[FOCAL-8] Use dynamic pixel mapping
Pixel mapping was restructured in
https://github.com/AliceO2Group/AliceO2/pull/11350
from a hard coded mapping corresponding to
the Testbeams before to a dynamic mapping where
the correspondece of FEE and Chip in HIC to a global
chip position in FOCAL is defined using config files
(and later CCDB). New mapping is used in the QC as
PixelMappingV1. Will be moved to PixelMapping once the
commit in QC is merged into master.