[QC-429] Trending tasks should use End Of Validity for point timestamps
If trended objects follow the new validity rules, they should use the end of validity as the timestamp of a data point, since the start will be usually the same (SOR).
I did not touch tasks which use creation time or current timestamp instead, since these will not be affected by the change.
FIT tasks have a way to select what kind of timestamps should be used, so I added a new option for "validUntil".