[QC-429] Assume nOrbitsPerTF = 32 until we retrieve it properly
The value from DataTakingContext is actually not set anywhere, so the default 128 was propagated, which is wrong at P2.
I am putting this in place, since I might not be ready with a proper solution before my holidays.
Removing this access will also let us remove the field from DataTakingContext, as agreed with David.