[QC-429] Get nOrbitsPerTF from CCDB in async QC
nOrbitsPerTF is equal to nHBFperTF by definition.

I hesitated to access it also in sync QC, because the access mechanism is not well testable and fails in ways I do not understand (complaints about token missing when I do have a token).
Fortunately, in sync QC this number is not crucial and so far it is going to be used only for reporting some statistics, thus I feel moderately comfortable with assuming it is 32.