[QC-1012] Allow TrendingTaskITSFhr to trend slightly desynchronized objects
The FHR Task generates objects specific to particular detector layers. Since detector layers are assigned to different subsets of FLPs, the merged objects
do not have common validity, as FLPs do not have QC cycles ideally in sync. Thus we have look for the objects a bit behind and forward from the trigger timestamp as well.