[QC-1023] Fix range string Y axis
Since the loop over the y slices is nested within
the loop over the x slices, the way the title was
constructed was leading to appending the new
y-range to previous y-ranges. Therefore later
slices contained all previous slices in y. The
x-range must be cached, and in the iteration of
the y-range for each slice a new title must be
constructed starting from the cached x-title.