[QC-1015] Do not trend slices if some objects are unavailable
We saw several crashes at P2 when one of the requested objects was not available.
In such case, `dataRetrieveVector->at(p)` calls would throw, since the vector
was not filled all requested data points. I think the most sane thing to do in such case,
is to avoid adding a new row in the TTree, otherwise we would have to fill it with 0,
which might appear as very misleading in some of the produced plots.