[QC-1036] Fix merging invalid validity (#2006)
As discovered, when a plot is empty it correctly gets invalid validity range (uint_max, 0). However, when we merge it with a correct non-empty object, the invalid range becomes the new range of the valid object (0, uint_max), which is wrong.
This PR makes invalid validity intervals be ignored when merging objects.