Allow for TFID == 0 in async QC timekeeping for testing purposes
As it has been demonstrated by a recent use case, some simulated data does not use correct TFID, having it as 0 instead. In such cases, we do not want to block the users from testing their sw, thus we allow for it and produce a warning.
We overwrite TFID to 1, so it does not break the algorithm which computes the data timespan (production data starts with TFID == 1)