[QC-1049][QC-1050] Allow to republish an object
This allows to publish an object with the same name again, which now overwrites the previous one.

This is needed for START STOP START in postprocessing, where a task might want to replace an object with a new one. It also makes the framework survive if a task attempts to register  the same object again, because the registration happens at PostProcessingInterface::initialize(), which we want invoke at each START.