[QC-686] Allow apply Reductors to CCDB data in TrendingTasks and alike
This commit brings the possibility to apply Reductors to CCDB objects.
To achieve that, the former Reductor (which assumes the expected type as TObject*) has become ReductorTObject, while ReductorConditionAny lets us retrieve any type which has a dictionary with retrieveConditionAny method.
For better separation and easier testability, I extracted that method into a separate ConditionAccess class.