Stop publishing all objects after PostProcessingInterface::finalize call
In postprocessing the task flow is always initialize()->update()->finalize(), where update() is done any number of times, including 0.
This allows us to safely stop publishing any objects after finalize() is called, because if the task will be reused in the next run, initialize() will be called first.
Hopefully this will fix many crashes during START->STOP->START related to the fact, that we were trying to start publishing objects which are already registered, but also deleted.