Avoid MonitorObjectCollection to grow in size
Currently the published objects are inserted in the
`MonitorObjectCollection` via the `TObjArray::Add()` method,
which always appends new elements at the end of the array.

When un-publishing objects, they are removed via the
`TObjArray::Remove()` method, which does not reduce the size of the
TObjArray. Instead, the removed slots are set to `nullptr`.

The proposed solution is to add a call to `TObjArray::Compress()`
after `TObjArray::Remove()`, which forces the removal of empty
slots from the array.