QC-1136 Publication Policy for ObjectsManager
This lets user ask for how long the registered object should be published.
By that, the user is made more aware of the expected lifetime of the object that they register,
which proved to be difficult for START STOP START in postprocessing.
SkeletonPostProcessing has been adapted to document the three main scenarios that may happen.

User tasks were adapted to use reasonable parameters.
On the occasion, a few problems leading to potentially faulty START STOP START were additionally fixed.