[Common] restored title in quality trending plots
The title of the plot was not shown because of the use of a dummy
graph to plot the axes.
The code has been simplified, removing the need of the dummy graph.
The output is visually identical as before, except that the plot title
is now also displayed, and the title of the horizontal axis is "time".