[MCH] add per-station distributions
The code adds some 1-D histograms that provide distributions of relevant
quantities (hit rates, cluster sizes and charge) separately for each
tracking station.
The plots are useful for diagnostics and comparison with reference data.

It also introduces a utility function to create a log10-spaced binning that
is suitable to initialize variable-bin-sized 1-D histograms for which the
X-axis should be displayed in log scale.