[Common] properly handle Sumw2 in THxRatio classes
In the original implementation the creation of the sum-of-weights
structure was forcibly enabled in the `init()` function, for both the
ratio histogram and the internal numerato/denominator ones, without
the possibility to disable it.
This doubles the size of the produced objects, because of the extra
memory needed to store the sum-of-weights structures.

The new implementation allows to disable the sum-of-weights on a
per-histogram basis, by calling `Sumw2(kFALSE)` immediately after
the instantiation of the histogram ratio. This applies to both the
ratio histogram and the internal numerator/denominator ones.

The sum-of-weights is still enabled by default in the `init()` function
of the histogram ratio, to match the previous behaviour and guarantee
that errors are correctly computed and propagated at each step.

The relevant changes are the following:
- added a `Bool_t mSumw2Enabled` member variable to keep track of wether or not the sum-of-weights structure should be updated in the histogram computations
- overridden the `void Sumw2(Bool_t flag)` virtual function to propagate the flag to the internal numerator/denominator histograms
- modified the `update()` function to call `T::Scale(norm, "nosw2")` when the sum-of-weights structure is not enabled (otherwise the `Scale()` function automatically creates the structure regardless of the user settings)