QC-1155 Allow for multiple graphs in one canvas in TrendingTask
This allows to draw more than one histogram or TGraph on one plot (canvas) in TrendingTask.
If more than one graph is used, a legend is automatically added to the plot.

The configuration scheme has been adapted accordingly to allow for multiple graphs.
The old way is still supported, but has been removed from documentation to encourage users to move to the new scheme.

There is also a possibility to select custom color palette expected, but we will need ROOT bump to 6.32 before this can be enabled.

Some refactoring is included as well.