O2-5072 TrendingTask reuses TTree across runs
In some cases we saw TrendingTask crash due to double deletion in a TTree when restarting a run. The stack trace was:
```
TNamed::~TNamed()
TLeafD::~TLeafD()
TObjArray::Delete(char const*)
TBranch::~TBranch()
TObjArray::Delete(char const*)
TTree::~TTree()
TTree::~TTree()
```
Since I was not able to find the root cause and the crash is not consistent, the best I can propose is to avoid deleting the TTree during restart.
Instead, we reuse the tree throughout the lifetime of the task.