QC-1202 Improve naming of canvas elements created by TrendingTask
As it turned out, some TPC Checks were relying on the default naming used by TTree::Draw when drawing the trends, i.e. "Graph" and "htemp".
When implementing QC-1155, we could not keep them default to allow for drawing a legend in case we had multiple trends on one canvas.

This commit changes the following:
- It adds an optional "name" field to graphs. If it's absent, "title" is used. If that is absent as well, the main "name" of the plot is used.
- In case that "htemp" is used by TTree::Draw only to draw the axes, title, etc, it is renamed to "background". In case that the plot is 1D, "htemp" is renamed to the plot "name"
- In case that errors are drawn, the object is renamed to "name" suffixed with "_errors"