[Common] run number specified in CcdbInspector queries
In the case of queries to the CCDB the metadata was completely empty, so
the latest version of the queried object was retrieved regardless of the
run number.
This leads to failures in retrieving the appropriate object in cases where
the same type of object is updated by multiple parallel runs, as it is the
case for example for the CTP configuration object. In such cases the object
from the wrong run might be retrieved, if it was created shortly after the
good one.

To avoid that this kind of corner cases leads to wrongly reported failures,
the run number is now explicitly set in the metadata used for the CCDB
queries, as it was already done for the QCDB ones.