[COMMON] allow retrieving latest QOs in QualityTask
Instead of retrieving the QOs based on the time stamp of the update trigger,
the latest version of the QOs is retrieved, even if their validities are
ending before the update time stamp.

A `maxObjectAgeSeconds` configuration parameter allows to consider only
QOs that were created not more than a given number of seconds before the
time at which the `update()` function was called.