[Common] added a bin-by-bin deviation comparator
The comparison based on the relative deviation between each bin of the
current and reference histograms.
A `maxAllowedBadBins` parameter allows to specify the maximum number of
bins above threshold for which the quality is still considered Good.
If X and/or Y ranges are specified, only the bins inside the ranges are
considered for the comparison.