[Common] improve layout of 1-D reference plot comparisons
The new layout puts the ratio plot above the two superimposed histograms,
and avoids the presence of a transparent pad above the histograms.
This restores the possibility to see the bin values by hovering the
mouse pointer over the histogram, as well as of zooming the
horizontal axis with the mouse.