[THRatio] fix update of histograms with bin labels
A direct copy of the axis objects from the numerator to the ratio histograms,
that was introduced in a recent commit, breaks the preservation of the
bin labels that are eventually set in the ratio histogram.

The code in this commit replaces the copy with an explicit update of the
number and limits of the bins, such that it works also if the numerator
is filled with `TH1::kCanRebin` set, and therefore the axis range can
change after the histogram creation.