[MCH] add rate and efficiency plots per DualSAMPA (#2533)
The new histograms are 1D lightweight versions of the 2D maps in
electronics coordinates, that are more suitable for saving moving windows
plots in async productions. The 1D versions provide the quantities averaged
over each DualSAMPA bard, while the 2D versions contain values on a
channel-by-channel basis.

By default the 1D histograms are enabled and the 2D are disabled.