[Common] allow access to the reference comparator plots in derived classes
The changes allow to use the ReferenceComparatorTask as a base class,
providing access to the reference comparison plots for the derived classes.

A typical use case would be a post-processing task that produces some
derived plots, which also need to compared with some reference.
With the current production code, this involves adding a dedicated
ReferenceComparatorTask in the workflow, that fetches the plots from the
post-processing task from the QCDB.

This commit allows to solve the above problem by deriving the post-processing
task from the ReferenceComparatorTask, such that the reference comparison
plots can be filled directly by the derived task, without going through
the QCDB.

This is achieved with the following two changes:
* addition of a function that exposes the reference comparison plots in the ReferenceComparatorTask interface
* storage of the pointer to the reference histogram in the ReferenceComparatorPlot object, such that it does not need to be passed to the `update()` method