[MID] fixed Digits check result in case of noisy boards
In the presence of noisy boards, the result of the check on the empty boards
was incorrectly ignored, leading in some cases to a Medium quality even
with a large number of empty boards being present.

The fix aggregates the qualities from the bad boards and empty boards
checks, and also displays messages from both chekers in case they both
detect problems.