Fix wrong warnings about bad flags assigned to good qualities (#2583)
Due to the order of processing in WorstOfAllAggregator, we could get cases when a bad Flag (coming from Bad quality) would be assigned to Good quality, while the quality itself would be replaced just after.
This commit reverses the order, which will silence the wrong warning.

The same code was also reused TrendCheck, thus I am fixing it also there.