MCH: added plots per cru link (#2592)
* [MCH] added averaging over SOLAR links

Added computation of average quantities over SOLAR links, similar to
what is already implemented for the averaging over Detection Elements

* [MCH] added helper functions for handling SOLAR indexes and histogram decorations

* [MCH] updated Pedestals Task histogram decorations

The changes are needed to sync with the updated helper code

* [MCH] added plots as function of SOLAR links

The commit adds plots with quantities averaged over SOLAR links,
similar to the existing ones as function of Detection Elements.

It also adds comparisons of DE and SOLAR plots with reference ones,
using the code from https://github.com/AliceO2Group/QualityControl/pull/2578

* [MCH] added checkers for SOLAR-based plots

The code that checks the SOLAR-based plots is similar to that for the
DE-based ones. The checkers also fill a summary plot with the overall
quality of each SOLAR link, which are then aggregared by a dedicated task.

* [MCH] added aggregator task for DE and SOLAR qualities

The task combines the summary qualities for each DE and SOLAR link,
and uploads a CCDB object with the list of bad DE and SOLAR IDs.

The CCDB objects will be used to automatically reconfigure the bad links
during the data taking.

* [MCH] fixed clang errors

* [MCH] added review suggestions