[QC-1297] Accept pure json as value instead of a string containing json.
A more elegant way would be to use a variant in the mCustomParameters data structure. However, it implies making quite many changes to check what we are dealing with in every existing method, and some of them would need heavy refactoring (e.g. find()). Given that the overhead of the proposed solution is minimal, I would stick with it.